scanner: Make ArtifactoryStorage's "read(Package, ScannerDetails)" the default
Move it to the interface so it can be shared with other implementations.
