+++ scanner/src/main/kotlin/ScanResultsStorage.kt
@@ -117,7 +117,40 @@ interface ScanResultsStorage {
      *
      * @return The [ScanResultContainer] matching the [id][Package.id] of [pkg] and the [scannerDetails].
      */
-    fun read(pkg: Package, scannerDetails: ScannerDetails): ScanResultContainer
+    fun read(pkg: Package, scannerDetails: ScannerDetails): ScanResultContainer {
+        val scanResults = read(pkg.id).results.toMutableList()
+
+        if (scanResults.isEmpty()) return ScanResultContainer(pkg.id, scanResults)
+
+        // Only keep scan results whose provenance information matches the package information.
+        scanResults.retainAll { it.provenance.matches(pkg) }
+        if (scanResults.isEmpty()) {
+            log.info {
+                "No stored scan results found for $pkg. The following entries with non-matching provenance have " +
+                        "been ignored: ${scanResults.map { it.provenance }}"
+            }
+            return ScanResultContainer(pkg.id, scanResults)
+        }
+
+        // Only keep scan results from compatible scanners.
+        scanResults.retainAll { scannerDetails.isCompatible(it.scanner) }
+        if (scanResults.isEmpty()) {
+            log.info {
+                "No stored scan results found for $scannerDetails. The following entries with incompatible scanners " +
+                        "have been ignored: ${scanResults.map { it.scanner }}"
+            }
+            return ScanResultContainer(pkg.id, scanResults)
+        }
+
+        log.info {
+            "Found ${scanResults.size} stored scan result(s) for ${pkg.id.toCoordinates()} that are compatible with " +
+                    "$scannerDetails."
+        }
+
+        val patchedScanResults = patchScanCodeLicenseRefs(scanResults)
+
+        return ScanResultContainer(pkg.id, patchedScanResults)
+    }
 
     /**
      * Add a [ScanResult] to the [ScanResultContainer] for this [id] and write it to the storage.
@@ -133,4 +166,28 @@ interface ScanResultsStorage {
      * List the [Identifier]s of all stored packages.
      */
     fun listPackages(): SortedSet<Identifier>
+
+    // TODO: Remove this code again once we migrated our scan result storage to contain the new "namespaced" license
+    // names for ScanCode.
+    fun patchScanCodeLicenseRefs(scanResults: List<ScanResult>) =
+        scanResults.map { result ->
+            if (result.scanner.name == "ScanCode") {
+                val findings = result.summary.licenseFindings.map { finding ->
+                    if (finding.license.startsWith("LicenseRef-") &&
+                        !finding.license.startsWith("LicenseRef-scancode-")
+                    ) {
+                        val suffix = finding.license.removePrefix("LicenseRef-")
+                        val license = "LicenseRef-scancode-$suffix"
+                        log.info { "Patched license name '${finding.license}' to '$license'." }
+                        finding.copy(license = license)
+                    } else {
+                        finding
+                    }
+                }
+
+                result.copy(summary = result.summary.copy(licenseFindings = findings.toSortedSet()))
+            } else {
+                result
+            }
+        }
 }
